requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.3
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.2.2
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-cors-headers==3.12.0
django-debug-toolbar==3.5.0
django-extra-fields==3.0.2
django-filter==22.1
django-mathfilters==0.1.2
django-multiselectfield==0.1.12
djangorestframework==3.13.1
duplicity==0.8.12.0
entrypoints==0.3
environs==8.0.0
fasteners==0.14.1
filelock==3.0.12
frozenlist==1.3.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
h11==0.13.0
httplib2==0.14.0
huepy==1.2.1
idna==2.8
imageio==2.19.5
imageio-ffmpeg==0.4.7
importlib-metadata==1.5.0
iniconfig==1.1.1
instabot==0.117.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
marshmallow==3.17.0
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
moviepy==1.0.3
multidict==6.0.2
netifaces==0.10.4
numpy==1.23.1
oauthlib==3.1.0
olefile==0.46
outcome==1.2.0
packaging==21.3
paramiko==2.6.0
paymeuz==0.2
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
pluggy==1.0.0
proglog==0.1.10
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.3
py==1.11.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.3.1
reportlab==3.5.34
requests==2.28.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.21.0
schedule==1.1.0
SecretStorage==2.3.1
selenium==4.3.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.10
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
whitenoise==6.1.0
wsproto==1.1.0
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
